# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Advanced Payment Processing API
  description: |
    Enterprise-grade payment processing system with distributed tracing, queue-based processing, 
    and comprehensive compliance features.
    
    ## Features
    - **Payment Processing**: Authorize, capture, void, refund operations
    - **Subscription Management**: Recurring payment lifecycle management  
    - **Webhook System**: Reliable event delivery with retry logic
    - **Database Operations**: Direct CRUD access to entities
    - **Queue Management**: Background job processing and monitoring
    - **Distributed Tracing**: End-to-end request correlation
    - **Compliance**: PCI DSS, GDPR, SOX compliant design
    
    ## Authentication
    All endpoints require API key authentication via `X-API-Key` header.
    
    ## Rate Limiting  
    - Standard endpoints: 1000 requests/hour
    - Payment endpoints: 500 requests/hour
    - Admin endpoints: 100 requests/hour
    
    ## Correlation IDs
    All responses include `X-Correlation-ID` header for request tracing.
    
  version: 3.0.0
  contact:
    name: Payment Processing API Support
    email: api-support@paymentprocessing.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.paymentprocessing.com/v3
    description: Production server
  - url: https://sandbox.paymentprocessing.com/v3  
    description: Sandbox server
  - url: http://localhost:3000
    description: Development server
security:
  - ApiKeyAuth: []
paths:
...
  /api/queues/stats:
    get:
      tags: [Queues]
      summary: Queue statistics
      description: Get queue processing statistics
      security:
        - ApiKeyAuth: [admin]
      responses:
        '200':
          description: Queue statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStatsResponse'
  /api/queues/info:
    get:
      tags: [Queues]
      summary: Queue system information
      description: Get queue system configuration and status
      security:
        - ApiKeyAuth: [admin]
      responses:
        '200':
          description: Queue information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueInfoResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    QueueStatsResponse:
      type: object
      properties:
        overview:
          type: object
          properties:
            totalJobs:
              type: integer
            completedJobs:
              type: integer
            failedJobs:
              type: integer
            activeJobs:
              type: integer
            waitingJobs:
              type: integer
        queues:
          type: object
          additionalProperties:
            type: object
            properties:
              completed:
                type: integer
              failed:
                type: integer
              waiting:
                type: integer
              active:
                type: integer
              delayed:
                type: integer
        performance:
          type: object
          properties:
            averageProcessingTime:
              type: number
            jobThroughput:
              type: number
            peakConcurrency:
              type: integer
        timeRange:
          type: string
          description: Time range for statistics (e.g., "24h")
    QueueInfoResponse:
      type: object
      properties:
        driver:
          type: string
          enum: [redis, memory]
        configuration:
          type: object
          properties:
            maxConcurrency:
              type: integer
            retryAttempts:
              type: integer
            retryDelay:
              type: integer
            jobTimeout:
              type: integer
        queues:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              workers:
                type: integer
              concurrency:
                type: integer
        uptime:
          type: integer
          description: Queue system uptime in seconds

    # Tracing Schemas
...